Fixing up the way COMMIT_RANGE is calculated
This PR arose from https://github.com/ubccr/xdmod-supremm/pull/263 in which the
style tests failed to pass. It seemed to think that the entire file was new and
was listing all of the style violations that were present, as opposed to only
taking into account the changes introduced by this PR. After further
investigation it was found that the problem was caused by travis related code
that was carried over during the migration to Shippable.

Specifically, the `if` statement on line 56:
`if [ -z "$COMMIT" -a -z "$COMMIT_RANGE" ]; then` failed due to the build not
having a COMMIT_RANGE which meant that it fell through and hit the following
statement `if [ -z "COMMIT_RANGE" ]; then ...` which meant that the COMMIT_RANGE
was incorrectly calculated as `"${COMMIT}^1")...${COMMIT}` thus causing the
style check to fail.